requirements.txt
Flask==3.1.2
pymongo==4.16.0
pandas==2.2.3
numpy==2.4.1
scikit-learn==1.5.2
sentence-transformers==5.2.0
transformers==4.57.6
torch==2.9.1
beautifulsoup4==4.14.3
webdriver-manager==4.0.2
playwright==1.57.0
requests==2.32.5
google-auth==2.49.1
python-dotenv==1.2.1
keybert==0.9.0
tqdm==4.67.1